feat(hm/dua)!: move the dua module from NixOS to home-manager
For the upcoming `eza` module, setting it in the system configuration was
causing issues with aliases set by Oh My Zsh. Let’s then set everything in the
home configuration so that aliases are not at risk of being overwritten.

BREAKING CHANGE: The NixOS dua module is moved to a home-manager module instead.
Please set `confkit.programs.dua.enable` in your home configuration instead of
the system one.